fix(frontend): store removed, helper added
